Change configuration variables in SpamFilter to SimpleProperties. Add H2 database and logging dependencies.
Changed member variables in SpamFilter to SimpleIntegerProperties, SimpleDoubleProperties, and SimpleBooleanProperties so that I can add listeners easily when I begin making the GUI.

Added a dependency for slf4j-simple, a logging library required by PircbotX. For some reason, I did not have it as a Maven depedency when I first started the project.

Added the H2 Database Engine library. User viewing data will be stored on an embedded database to make storage easy.